requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
aspy.yaml==1.3.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
blis==0.2.4
certifi==2019.6.16
cfgv==2.0.1
chardet==3.0.4
cymem==2.0.2
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
h11==0.8.1
h2==3.1.0
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
identify==1.4.5
idna==2.8
importlib-metadata==0.19
more-itertools==7.1.0
multidict==4.5.2
murmurhash==1.0.2
nltk==3.4.5
nodeenv==1.3.3
numpy==1.17.0
packaging==19.0
plac==0.9.6
pluggy==0.12.0
pre-commit==1.18.0
preshed==2.0.1
py==1.8.0
pyparsing==2.4.0
pysqlite3==0.2.1
pytest==5.0.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytest-asyncio==0.10.0
PyYAML==5.1.2
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.6.2
six==1.12.0
spacy==2.1.7
srsly==0.0.7
thinc==7.0.8
toml==0.10.0
tqdm==4.32.2
ujson==1.35
urllib3==1.25.3
uvloop==0.12.2
virtualenv==16.7.2
wasabi==0.2.2
wcwidth==0.1.7
websockets==6.0
yapf==0.28.0
yarl==1.3.0
zipp==0.5.2
